Moved bans.js to the models directory
- Moved bans.js
- Removed the bans import in `user.js` and `pilotSettingsController.js` files.
- Corrected the reference for the bans controller in `adminBansController.js` and `users.js`
