+++ controllers/pilotSettingsController.js
@@ -1,9 +1,3 @@
-var path = require('path');
-var setup = require('../setup.js');
-var bans = require('../bans.js')(setup);
-var esi = require('eve-swagger');
-const log = require('../logger.js')(module);
-
 //Render Pilot Settings Page
 exports.index = function(req, res) {
     if (req.isAuthenticated()) {
